# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Zylo Errand Marketplace API
  description: |
    Complete REST API for the Zylo mobile application - An Indrive-style errand marketplace platform.

    ## Features
    - JWT-based authentication
    - Errand marketplace with escrow payments
    - Real-time chat messaging
    - Negotiation/counter-offer system
    - Rating and review system
    - Tier-based fee structure with auto-promotion
    - Complete wallet system with transaction ledger

    ## Authentication
    Most endpoints require JWT authentication. Click the **Authorize** button and enter your JWT token.

    ## Testing
    1. Sign up or login to get a JWT token
    2. Click **Authorize** and paste the token
    3. Try out any endpoint using the "Try it out" button

  version: 1.0.0
  contact:
    name: Zylo API Support
    email: support@zylo.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.zylo.app
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/errands:
    get:
      tags:
        - Errands
      summary: List all errands
      description: Get a list of errands with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, assigned, in_progress, completed, cancelled]
          description: Filter by errand status
        - name: category
          in: query
          schema:
            type: string
            enum: [fuel_energy, courier_delivery, office_work, custom]
          description: Filter by category
        - name: sender_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by sender user ID
        - name: errander_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by errander user ID
      responses:
        '200':
          description: Errands retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          errands:
                            type: array
                            items:
                              $ref: '#/components/schemas/Errand'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Errands
      summary: Create a new errand
      description: Post a new errand to the marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrandCreate'
      responses:
        '201':
          description: Errand created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          errand:
                            $ref: '#/components/schemas/Errand'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token from login/signup
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        error:
          type: string
...
    Location:
      type: object
      required:
        - latitude
        - longitude
        - address
      properties:
        latitude:
          type: number
          example: 6.5244
        longitude:
          type: number
          example: 3.3792
        address:
          type: string
          example: 123 Main St, Lagos
    Errand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        errander_id:
          type: string
          format: uuid
        category:
          type: string
          enum: [fuel_energy, courier_delivery, office_work, custom]
        title:
          type: string
        description:
          type: string
        budget:
          type: number
        pickup_location:
          $ref: '#/components/schemas/Location'
        destination_location:
          $ref: '#/components/schemas/Location'
        status:
          type: string
          enum: [open, assigned, in_progress, completed, cancelled]
        image_urls:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    ErrandCreate:
      type: object
      required:
        - category
        - title
        - description
        - budget
        - destination_location
      properties:
        category:
          type: string
          enum: [fuel_energy, courier_delivery, office_work, custom]
          example: courier_delivery
        title:
          type: string
          example: Deliver package to Ikeja
        description:
          type: string
          example: Need package delivered urgently
        budget:
          type: number
          example: 5000
        pickup_location:
          $ref: '#/components/schemas/Location'
        destination_location:
          $ref: '#/components/schemas/Location'
        image_urls:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid request
                  error:
                    type: string
                    example: Missing required fields
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Unauthorized
                  error:
                    type: string
                    example: User not authenticated
...